+++ spec/lib/dawn/codesake_knowledgebase_spec.rb
@@ -718,4 +718,9 @@ end
   sc.should_not   be_nil
   sc.class.should == Codesake::Dawn::Kb::CVE_2013_2119
 end
+  it "must have test for CVE-2013-7086" do
+  sc = kb.find("CVE-2013-7086")
+  sc.should_not   be_nil
+  sc.class.should == Codesake::Dawn::Kb::CVE_2013_7086
+end
 end